extract range to a component
