+++ components/definition/populations/NetPyNEPopulation.js
@@ -13,14 +13,13 @@ import FontIcon from 'material-ui/FontIcon';
 import clone from 'lodash.clone';
 import Utils from '../../../Utils';
 import NetPyNEField from '../../general/NetPyNEField';
-import AdapterComponent from '../../general/AdapterComponent';
 import DimensionsComponent from './Dimensions';
+import RangeComponent from './Range';
 
 var PythonControlledCapability = require('../../../../../js/communication/geppettoJupyter/PythonControlledCapability');
 var PythonControlledTextField = PythonControlledCapability.createPythonControlledControl(TextField);
 var PythonControlledSelectField = PythonControlledCapability.createPythonControlledControl(SelectField);
 var PythonControlledAutoComplete = PythonControlledCapability.createPythonControlledControl(AutoComplete);
-var PythonControlledAdapterComponent = PythonControlledCapability.createPythonControlledControl(AdapterComponent);
 
 
 export default class NetPyNEPopulation extends React.Component {
@@ -34,13 +33,6 @@ export default class NetPyNEPopulation extends React.Component {
       selectedIndex: 0,
       sectionId: "General"
     };
-
-    
-    this.ranges = [
-      { value: 'xRange', stateVariable: 'rangeTypeX' }, { value: 'xnormRange', stateVariable: 'rangeTypeX' },
-      { value: 'yRange', stateVariable: 'rangeTypeY' }, { value: 'ynormRange', stateVariable: 'rangeTypeY' },
-      { value: 'zRange', stateVariable: 'rangeTypeZ' }, { value: 'znormRange', stateVariable: 'rangeTypeZ' }
-    ]
   }
 
   componentWillReceiveProps(nextProps) {
@@ -56,7 +48,6 @@ export default class NetPyNEPopulation extends React.Component {
           return result;
         }, {});
       this.setState(newState);
-      this.updateLayout();
     }
   }
 
@@ -90,30 +81,6 @@ export default class NetPyNEPopulation extends React.Component {
 
   }
 
-  updateLayout() {
-
-    //Range Type X
-    const getRange = (value, stateVariable) => {
-      Utils
-        .sendPythonMessage("'" + value + "' in netParams.popParams['" + this.state.model.name + "']")
-        .then((response) => {
-          if (response) {
-            var newState = {};
-            newState[stateVariable] = value;
-            this.setState(newState)
-          }
-        });
-    }
-
-    this.ranges.forEach((range) => { getRange(range.value, range.stateVariable) })
-
-
-  }
-
-  componentDidMount() {
-    this.updateLayout();
-  }
-
   getModelParameters = () => {
     var select = (index, sectionId) => this.setState({ selectedIndex: index, sectionId: sectionId })
 
@@ -129,7 +96,12 @@ export default class NetPyNEPopulation extends React.Component {
   }
 
   shouldComponentUpdate(nextProps, nextState) {
-    return this.state.model == undefined || this.state.currentName != nextState.currentName || this.state.cellModelFields != nextState.cellModelFields || this.state.model != nextState.model || this.state.dimension != nextState.dimension || this.state.sectionId != nextState.sectionId || this.state.selectedIndex != nextState.selectedIndex || this.state.rangeTypeX != nextState.rangeTypeX || this.state.rangeTypeY != nextState.rangeTypeY || this.state.rangeTypeZ != nextState.rangeTypeZ;
+    return this.state.model == undefined ||
+      this.state.currentName != nextState.currentName ||
+      this.state.cellModelFields != nextState.cellModelFields ||
+      this.state.model != nextState.model ||
+      this.state.sectionId != nextState.sectionId ||
+      this.state.selectedIndex != nextState.selectedIndex;
   }
 
   handleRenameChange = (event) => {
@@ -156,7 +128,6 @@ export default class NetPyNEPopulation extends React.Component {
     if (this.state.sectionId == "General") {
       var content =
         <div>
-
           <TextField
             onChange={this.handleRenameChange}
             value={this.state.currentName}
@@ -180,110 +151,10 @@ export default class NetPyNEPopulation extends React.Component {
           </NetPyNEField>
 
           <DimensionsComponent model={this.state.model} />
-          
         </div>
     }
     else if (this.state.sectionId == "SpatialDistribution") {
-      var content =
-        <div>
-          <NetPyNEField id={"netParams.popParams.xRange"} >
-            <SelectField
-              floatingLabelText="Range type"
-              value={this.state.rangeTypeX}
-              onChange={(event, index, value) => this.setState({ rangeTypeX: value })}
-            >
-              <MenuItem key="xRange" value="xRange" primaryText="Absolute" />
-              <MenuItem key="xnormRange" value="xnormRange" primaryText="Normalized" />
-            </SelectField>
-          </NetPyNEField>
-          {(this.state.rangeTypeX != undefined) ?
-            <div className={"netpyneRightField"}>
-              <PythonControlledAdapterComponent
-                model={"netParams.popParams['" + this.state.model.name + "']['" + this.state.rangeTypeX + "']"}
-                convertToPython={(state) => {
-                  if (state.minXAxis != undefined && state.maxXAxis != undefined) {
-                    return "[" + state.minXAxis + "," + state.maxXAxis + "]";
-                  }
-                }}
-                convertFromPython={(prevProps, prevState, value) => {
-                  if (value != undefined && prevProps.value != value) {
-                    var arrayValue = JSON.parse(value);
-                    return { minXAxis: arrayValue[0], maxXAxis: arrayValue[1] };
-                  }
-                }}
-              >
-                <TextField floatingLabelText="Min x-axis" id="minXAxis" />
-                <TextField floatingLabelText="Max x-axis" id="maxXAxis" />
-              </PythonControlledAdapterComponent>
-            </div>
-            : null}
-          <br />
-          <NetPyNEField id={"netParams.popParams.yRange"} >
-            <SelectField
-              floatingLabelText="Range type"
-              value={this.state.rangeTypeY}
-              onChange={(event, index, value) => this.setState({ rangeTypeY: value })}
-            >
-              <MenuItem key="yRange" value="yRange" primaryText="Absolute" />
-              <MenuItem key="ynormRange" value="ynormRange" primaryText="Normalized" />
-            </SelectField>
-          </NetPyNEField>
-          {(this.state.rangeTypeY != undefined) ?
-            <div className={"netpyneRightField"}>
-              <PythonControlledAdapterComponent
-                model={"netParams.popParams['" + this.state.model.name + "']['" + this.state.rangeTypeY + "']"}
-                convertToPython={(state) => {
-                  if (state.minYAxis != undefined && state.maxYAxis != undefined) {
-                    return "[" + state.minYAxis + "," + state.maxYAxis + "]";
-                  }
-                }}
-                convertFromPython={(prevProps, prevState, value) => {
-                  if (value != undefined && prevProps.value != value) {
-                    var arrayValue = JSON.parse(value);
-                    return { minYAxis: arrayValue[0], maxYAxis: arrayValue[1] };
-                  }
-                }}
-              >
-                <TextField floatingLabelText="Min y-axis" id="minYAxis" />
-                <TextField floatingLabelText="Max y-axis" id="maxYAxis" />
-              </PythonControlledAdapterComponent>
-            </div>
-            : null}
-          <br />
-
-          <NetPyNEField id={"netParams.popParams.zRange"} >
-            <SelectField
-              floatingLabelText="Range type"
-              id={"netParams.popParams." + this.state.rangeTypeZ}
-              value={this.state.rangeTypeZ}
-              onChange={(event, index, value) => this.setState({ rangeTypeZ: value })}
-            >
-              <MenuItem key="zRange" value="zRange" primaryText="Absolute" />
-              <MenuItem key="znormRange" value="znormRange" primaryText="Normalized" />
-            </SelectField>
-          </NetPyNEField>
-          {(this.state.rangeTypeZ != undefined) ?
-            <div className={"netpyneRightField"}>
-              <PythonControlledAdapterComponent
-                model={"netParams.popParams['" + this.state.model.name + "']['" + this.state.rangeTypeZ + "']"}
-                convertToPython={(state) => {
-                  if (state.minZAxis != undefined && state.maxZAxis != undefined) {
-                    return "[" + state.minZAxis + "," + state.maxZAxis + "]";
-                  }
-                }}
-                convertFromPython={(prevProps, prevState, value) => {
-                  if (value != undefined && prevProps.value != value) {
-                    var arrayValue = JSON.parse(value);
-                    return { minZAxis: arrayValue[0], maxZAxis: arrayValue[1] };
-                  }
-                }}
-              >
-                <TextField floatingLabelText="Min z-axis" id="minZAxis" />
-                <TextField floatingLabelText="Max z-axis" id="maxZAxis" />
-              </PythonControlledAdapterComponent>
-            </div>
-            : null}
-        </div>
+      var content = <RangeComponent model={this.state.model} />
     }
     else if (this.state.sectionId == "CellList") {
       var content = <div>We should replicate population parameters</div>
